backend/requirements.txt
flask
flask-cors
flask-session
boto3
pillow
psycopg2-binary
numpy
torch
transformers
scikit-learn
python-dotenv